+++ src/fileio.h
@@ -9,9 +9,9 @@
 #include <string>
 #include "core/enum_type.hpp"
 
-void FioSeekTo(uint32 pos, int mode);
-void FioSeekToFile(uint8 slot, uint32 pos);
-uint32 FioGetPos();
+void FioSeekTo(size_t pos, int mode);
+void FioSeekToFile(uint8 slot, size_t pos);
+size_t FioGetPos();
 const char *FioGetFilename(uint8 slot);
 byte FioReadByte();
 uint16 FioReadWord();
